Issue #26426 Drafting a solution for the retrieval of runners that belong to subgroups for the owner of the parent group
Issue #26426 : Using the owned_groups with self_and_descendants to get
all the groups that a user owns even per inheritance of ownership.
Fixing the test case so that it behaves as expected.

Issue #26426 : No need to have two separate cases for group_runners and
subgroup_runners since the subgroups inherit the ownership from the
groups.

Issue #26426 : Improving description of the example.

Issue #26426 : As discussed, get rid of the example extraction and
inline it in the test case.

Issue #26426 : As discussed, adding a test case for inclusion of runners
that belong to groups in which the user is maintainer (and developer).

issue #26426 : Adding reporter and guest to shared examples for
completion.

Issue #26426 : Adding multi-owner, group owner and project owner as
shared examples.

Issue #26426 : Refining the case of group owner.

Issue #26426 : Making all let imperative. Reviewing the cases so that
they're complete.

Issue #26426 : Improving wording on the descriptions of the test. Adding
completion cases tests.