Move snippet commit ousitde update transaction
At the moment, when we update the snippet, we
also commit the information to the repository.
We do this inside a db transaction.

Nevertheless, because of this, some info is
not available to the different thread handling
the commit action.

This creates a race condition with the storage.

In this commit we move the commit action
outside the snippet db transaction.