Skip generating smaller designs for TIFFs
Our ability to enable the feature to generate smaller design images
https://gitlab.com/gitlab-org/gitlab/-/issues/207351 was blocked when it
was discovered that our servers did not have the ability to generate
JPG or TIFF https://gitlab.com/gitlab-org/gitlab/-/issues/207740.

The developer who is working on allowing our infrastructure to resize
images has enabled JPGs to be resized in production, however TIFF
support may take a couple more weeks
https://gitlab.com/gitlab-org/gitlab/-/issues/207740#note_316655702.

We've decided to make a change to skip resizing TIFFs for now, and then
go ahead and enable the resizing feature, rather than to hold it up any
longer. We can then do a follow-up issue to enable TIFF when we have it.

https://gitlab.com/gitlab-org/gitlab/-/issues/213468