Stringify Sidekiq job args in exception logs
If these logs are sent to Elasticsearch, it will not be able to process
nested object fields, as this causes a type mismatch with scalar
elements in the same array across log lines.

This was done in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26267, but that
change did not apply to Sidekiq exceptions.

We also move all Sidekiq agument formatting into the JSON formatter.
This puts the formatting of job logs in one place and avoids the
pitfalls of altering job arguments in the actual payload.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/213639