Ensure NPM dev fixtures have unique name
NPM package names must be unique across a namespace. Previously the
development seed would fail if 2 of the 5 random projects selected to be
seeded were in the same group because the same package name would be
reused to seed the second project. To fix this, we append a random hex
string so that this conflict is unlikely to happen.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/213786