Delete a monkey-patch of Elasticsearch::Model
Class level caching, classes being created in tests and then stored in a
global registry of classes leads to bad results. The deleted test is
messing with a class level cache and creating classes, the previous
search method we switched out was using was looking through a global
registry of classes.