Add event to handle closing all board sidebars
This was made to avoid having to _manually_ close all sidebars (to open
a new one for example), and to prevent having two or more sidebars open
at the same time.

Each of the sidebars components are listening to a 'sidebars.closeAll'
event that has an optional `omit` property to avoid closing a sidebar
if not needed, for example, when browsing through issues.