Create group import UI
We want users to be able to create groups from imports in the UI as well
as the API. Create the user interface to achieve this.

For now, this is a copy of the layout and functionality of project
import but it is intended that we will iterate on this with a more
streamlined import process.

Relates to: https://gitlab.com/gitlab-org/gitlab/-/issues/211807