Add status page url to setting form in UI
Adds a form field to the status page setting form
under Settings > Operations > Status page. This
will enable the addition of features which depend
on the GitLab project knowing the url of the status
page to communicate to the user and make it easy
to navigate to the status page.