Replace isContentLoaded getter with state property
When a user submits empty content in the Static Site Editor,
the publish toolbar dissappears because the originalContent
property becomes empty. The isContentLoaded property depends
on the originalContent property to determine if there is
content to display.

This change decouples detecting that content is loaded from
detecting changes in the content