Better Geo Out of Date Errors
Currently when a Geo Node gets out of sync
the errors are scattered all around the card.

This is hard to follow for the user as the
location of the errors isn't always consistent.

This MR cleans up a lot of the current error UI
and consildates it.

The time since last synced is also now always
visible in the main section of the card.