Fix Status Page active status and remove unused StatusPageSetting
In
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/28966
we've removed the beta feature flag to make Status Page
generally available.

Thereafter, in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/29655
we've moved StatusPageSetting to StatusPage::ProjectSetting.

Apparently, we've missed to remove this model and just added
`StatusPage::ProjectSetting`.

Addionally, due to a wrong resolution of a merge conflict we've
added back the beta feature flag to the new model
StatusPage::ProjectSetting. This made the Status Page setting page
behave weirdly: It showed up but users weren't able to activate.

This commits fixes both issues:
* Remove unused StatusPageSetting model
* Remove beta feature flag from StatusPage::ProjectSetting