Clean up refresh fix for cancel automatic merge
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26232, we fixed
the merge request widget not updating properly when clicking the `Cancel
automatic merge` button.

This fix was a bit of hack: it would force the frontend to refresh the
merge request widget instead of relying on the output of the
`/cancel_auto_merge` response. This response returned incomplete
information when we split the widget information in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/15045 to return
cached and uncached data.

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/29757, we fixed
the `/cancel_auto_merge` response to return all the relevant data, so we
can remove this workaround.