Switch RedisCacheable to use Redis cache instance
This used the 'shared state' Redis due to an understandable
misunderstanding about the purpose of that. The shared state Redis is
for persistent data, and the cache Redis is for caching short-lived
data.

Despite the names, both the cache and the shared state instance are
available to all ways the Rails application is deployed (web nodes and
background job nodes in particular).