Create UI for Group export in Group Settings
We want to be able to export group data from the UI like we currently
can with projects. Update the Group settings page to have link to
export the group data and then download it.

Add documentation describing how to complete this process from the UI.

Relates to issue: https://gitlab.com/gitlab-org/gitlab/-/issues/211805