Don't create pipelines when preparing RCs
In the days leading up to a monthly release, Release Tools will
automatically create a new release candidate every day. As part of this
process we merge the latest changes deployed to production into the
latest stable branch. Since we use merge requests for this (as our API
does not let you merge a branch directly), this will create a pipeline.
To deal with this we implemented some code to cancel pipelines before
merging these merge requests.

Sadly this proved unreliable, as sometimes pipelines are not yet
available after creating the merge request. If we then try to merge the
merge request this may fail if there are pending pipelines.

In this commit we change the CI config to not create a pipeline in the
first place for these merge requests. This is the easiest and most
reliable way of letting Release Tools merge its own merge requests,
without the need for any logic that waits for and cancels any running
pipelines.

This fixes https://gitlab.com/gitlab-org/release-tools/-/issues/423