Fix minor spacing issue at Snippet blob viewer
Prior to this change,
The padding between the two button groups in the header of a blob view for snippets is an HTML whitespace, so about 3.95px. According to Pajamas, spacing for this should actually be 8px.

How issue is resloved ?,
 After enabling Vue Snippet, the horizontal margin was set using mx-2 class to button group. the extra inline spacing was removed by making the parent class as flexbox.

Changes made in following files :
blob_header_viewer_switcher.vue - adding right padding
blob_header.vue - making the header element as flex