Add error feedback for non-existent design
Previously, when opening Design View, if the design
did not exist (i.e. bad url, wrong version) we
would not handle this case and the loading
spinner would endlessly display.

This commit adds better guards to extractDesign()
util to prevent an exception being thrown in this case,
and handles it explicitly by redirecting user and using
createFlash.

In addition, a few full stops are added to error messages
for consistency.