Change logic to find the current license
At the moment the last uploaded license is considered as the current
license. But when uploading a license which will start in the future it
is not supposed to be active right away. Therefore we will now find the
last uploaded license that has already started  and use it as the
current one.