Fix 500 error accessing restricted visibility levels
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/22522 introduced a
regression that caused a 500 error if the
`application_settings.restricted_visibility_levels` column were `nil`.
Since there is no NOT NULL constraint, there's no guarantee that it is
populated with the proper serialized YAML data.

To fix this specific problem, we can account for `nil` values, but we
should fix this at the database level.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/215552