Remove paid signup flow feature flag
Removes the paid sign up flow feature flag, which was rolled out
as an experiment. The experiment was successful and already used
for 100% of the users.

The new paid sign up flow is handling subscription purchases via
GitLab by communicating to a customers.gitlab.com API instead of
redirecting users to customers.gitlab.com