Fix broken tests affected by current license logic
With the new logic to find the current license some tests broke due to
using Timecop. The current code was not able to find a current license
because the usage of a past date was used for Timecop which was
before the license's start date. Therefore we set the license factory's
start date to 1970-01-01.