Mark ProcessCommitWorker as idempotent
The ProcessCommitWorker processed 1 million "duplicate"
jobs the past 7 days, and spent 120 hours on that.

Duplicate jobs are jobs that get scheduled when there
is already a job in the queue for the same worker with
the same arguments.

If the job was (marked as) idempotent,
we would be able to deduplicate those jobs automatically
when they get scheduled.