Fix file descriptor leak with Capybara
When Capybara is used in a debugger in feature specs, developers will
often see "Too many open files". As described in
https://github.com/teamcapybara/capybara/issues/2322, it appears that
this may be an issue with WebMock, which stops `Net::HTTP.start` from
connecting. If you start a connection and issue multiple requests with
it, WebMock will create a new connection per request.

Setting `net_http_connect_on_start: true` appears to fix the problem:
https://github.com/bblimke/webmock/blob/master/README.md

Many specs depending on the behavior of WebMock stubbing
Net::HTTP.start, so use the WebMock default behavior in most cases and
enable it only for feature specs. Kubeclient needs an exception because
as well, so add a :kubeclient label for this.