Add a historical count method to Vulnerability
Adds Vulnerability.count_by_day_and_severity. This method takes an
integer and returns vulnerability counts grouped by severity for that
number of days before the current day. It does not return rows when
there are no vulnerabilities for a day and severity.

The count does not include dismissed or resolved vulnerabilities.

This work is the first step towards moving the group security dashboard
history chart to first class vulnerabilities, and hopefully improving
its performance.

https://gitlab.com/gitlab-org/gitlab/-/issues/201792