Generalize Geo Filter Bar
As part of the effort of:
https://gitlab.com/gitlab-org/gitlab/-/issues/213219

We are combinding the views and navigation between
Geo Replicables.

To do this, we need a unified way to filter/search
the data between all views.

This MR pulls out the filter bar into its own partial,
so that it can be used for all replicables that
live in HAML.