Add Vale test for warning-level substitutions
There are some word choices we want to disallow in the future, but
the cleanup to get rid of them requires more time than we can spend
on the entire docset at this time. A moderate approach would be to
create a warning-level (as opposed to error-level) substitution test
for these word pairs, so we can start cleaning out the informal
versions of these words over time. If we focus on cleaning up all
instances of a word and occurrences hit zero, we can move a word
from the warnings file to the errors file.