Move all logic out of views
This moves even the path generation and permission checking logic out of
the views, making the controller responsible for packing up all the data
needed for display. This has the advantage of allowing view tests to use
the :build strategy, since no persistence is needed.

The @related_branches attribute is combined with @branch_info, so we
have just a single assigment.