Use MR SOURCE_BRANCH_NAME to prevent RC pipelines
When creating a merge request to merge release-tools/X into a stable
branch, using CI_COMMIT_BRANCH still results in a detached pipeline
being created, likely due to this variable not being set in that case.
To work around this we have to use CI_MERGE_REQUEST_SOURCE_BRANCH_NAME
instead, which is always set to the value we'd expect.

This fixes https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/822