Ensure wiki event creation is idempotent
This ensures that we don't create duplicate events due to event handlers
being run twice, once in the services and once in the post-receive
handlers.

To achieve this wiki page metadata objects have their update-at
attribute set to the most recent commit time, so that we can know if an
event for a given page state has already been applied. In the case of
deletes, where there is no most-recent-commit, we look to see if the
most recent event for a page is a deletion.

In future work, we would probably want to use commit shas here, but this
allows this feature to function without DB changes.