Add subscription banner to group/subgroup pages
The subscription banner will be displayed on group/subgroup pages, as
well as project pages for subgroups.

The most significant change here is looking up to the root_ancestor when
checking for a gitlab_subscription. The decorated_subscription as well
as the subscription_message now check project  first, and group second.

Finishes: https://gitlab.com/gitlab-org/growth/product/-/issues/1550