Upgrade json gem, handle differences
This upgrades the `json` gem but by default maintains the old way
of handling parsing as this is relied upon in a few places. This
also adds support for a `legacy_mode` flag when parsing, which will
be added manually to areas that require this in future. Currently
the legacy_mode is the default.