Move policies for group wikis into EE
This wasn't decided yet when we started working on it, but the feature
also hasn't been released yet so there shouldn't be any implications
for this move.

This also refactors `wiki_policies_shared_examples.rb` to verify the
permissions in all situations, and fixes a bug where
`:download_wiki_code` should be available on public or internal groups,
even if the user is not logged in or a member.