Add tests for passing nil values in epics and packages
In https://gitlab.com/gitlab-org/gitlab/-/issues/215936, we found that
upgrading to Grape v1.3.2 caused a regression in our QA suite because
passing a `nil` value for an Array type would be passed in as a `nil`
value instead of being coerced to an empty Array. We had to revert the
upgrade as a result. To ensure that `nil` values are handled properly on
the backend, add tests for this.