Decouple partial clone config
Previously partial clone configs is coupled with receive_max_input_size
check. This make sense when you think of partial clone as a replacement
for Git LFS but in truth, the 2 features are separated and should not be
 decoupled.

In self-managing instances, `receive_max_input_size` could be set to 0
while users could still wish to have partial clone feature enabled.

References:
- [Original discussion](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/16850#note_313323533)
- [Tracking issue](https://gitlab.com/gitlab-org/gitlab/-/issues/212926)