Set MergeRequestDiff#external_diff_store in more cases
Previously, we skipped setting this value on import, and only made a
change if the external_diff column had changed for a row. However, this
turned out to be counter-productive.

On import, MR diff files are not created through this mechanism anyway,
and sometimes we can update an MR diff while not changing the value of
the external_diff column (e.g. on migration). Just setting this should
be safe in both cases.