Allow administrator to export audit events to CSV
This change allows administrator to export audit events to CSV file. The
maximum file size is capped to 15Mb since it can fit nicely as an email
attachment and is a common limit across other CSV reports.

We decided not to cap by number of records (i.e 10k audit events) due to
some memory concerns since it requires 10k of records being fetched as
part of the preloading step.