Add more Praefect to scrape config
Each Praefect votes on the promotion of Gitaly nodes to primary, and
there should not be a single point of failure. Three Praefects is
preferred to two to prevent a tie. The scrape config should also
reflect this to make it clear that the Praefect load balancer IP is
not the correct address for this configuration.