Clear Redis-queues on Sidekiq::Testing.inline!
Since in specs we run though all the middleware for Sidekiq jobs, and
that Sidekiq middleware writes to Redis for job deduplication we
should clear that Redis before and after each spec to make sure that
the idempotency keys don't bleed into other specs.

The client middleware for job-deduplication writes a key to redis. The
server middleware removes it again before running the job. If the
spec were to get interrupted between these 2 steps, then the key could
remain in Redis causing failures on the next run in development.