Introduce check for future_dated licenses
With the prevention of future-dated licenses being considered as the
current one, there needs to be a new check for a new page when the first
uploaded license is a future-dated license. In this case a new page
will be displayed since there's no active license and we would normally
show the missing license page. This merge request will introduce the
check for the new page.