Log when container registry permissions are denied
If a user does not have access to pull or push to a container registry,
the /jwt/auth endpoint can still return a 200 success but the registry
may reject the request due to insufficient access. To help track down
intermittent failures, we now log a warning in `auth.log` if the
requested permissions don't match the granted permissions.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/23644