Avoid loading workers before setting Sidekiq defaults
Not a current bug, but it's possible that Sidekiq workers
are referenced before having a chance to load its defaults.

It means that it could run a worker with a wrong `retry`
default, for instance.

Here we solve that by loading it at the
`config/initializers_before_autoloader`, which is a
initializer that loads before autoloading happens.