Implement Vulnerabilities::Stats model
This model will be used to calculate the statistics for vulnerable
projects. Instead of loading these stats per project, we can just use
the builtin eager loading feature of ActiveRecord to load all the stats
for all the projects we are dealing with, which will solve the N+1
query problem we are having at the moment.