Rename Praefect HA to Gitaly Cluster
High Availability is property of a system, and is confusing when used
as the name of a feature. Particularly it is confusing given there are
multiple aspects to making a GitLab instance HA.

Additionally, the work we are doing on Gitaly and Praefect is not
exclusively for High Availability. A clustered configuration will
provide performance improvements in the future.