Add import in progress screen for group imports
Add an import in progress screen when a group has an ongoing import. For
now add functionality that matches the existing functionality for
project import.

Ensure that the import state is created when the import job is scheduled
so that this page immediately displays after the user creates the group,
but the import has not actually started yet.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/211808
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/31731